feat: automatically translate failed requests to localhost to docker.host.internal
When running inside docker, if a request to localhost fails with ECONNREFUSED retry
the request to docker.host.internal, returning the response if it succeeds but returning
the original error if it fails.

This is enabled by default, but only loaded when actually running inside docker, and
can be disabled with a new setting that defaults to true.