feat: add support for persisted operations (#249)
Add support to upload persisted operations using the CLI to the controlplane, saving them in S3 storage. From the router, add support for retrieving persisted operations from the Cosmo CDN when the client sends a request indicating a PO should be used.

Persisted operations are for now scoped to the org/federated graph/client-id and can never be deleted.

Additionally, persisted operation keys are always derived using the sha256 hash of the operation body. This allow us to use a content-addressable URL scheme that can be cached forever.

Co-authored-by: JivusAyrus <suvijsurya76@gmail.com>
Co-authored-by: Dustin Deus <deusdustin@gmail.com>