refactor: make the main HTTP router not dependent on the active GraphQL server
By decoupling the top level router's HTTP router and the currently active GraphQL server, we enable the router to be decoupled from running the HTTP server.

This separates concerns between configuring and executing the HTTP handler from running the server, and enables consumers to run their own HTTP handler while the internal logic can hot-swap when a new graph configuration is available.

Some consumers may already have an HTTP server running and simply want the router to be attached on a given path. This is the first PR along the path of enabling this capability.

This PR is purely an internal refactor. There are zero logical or API changes.