Tweak data migration to maintain table name traces_mv
The migration is not consistent with the overview in the readme “ We need to drop both the traces_mv materialized view and traces table, recreate and repopulate them. We do the repopulation in hourly batches.” as traces_mv is renamed to temp_traces_mv. This PR tweaks the code to make it consistent with the readme.