fix: generate plans for subscriptions (#2097)
Implement plans for subscriptions. Represent them the same way as Apollo does:

    QueryPlan {
      Subscription {
        Primary: {
          Fetch(service: "messages") {
            {
              messageInConversation(id: $id) {
                id
                text
                sentTime
              }
            }
          }
        },
        Rest: {
          Fetch(service: "employees") {
            {
              ...
            }
          }
        },
      }
    }

"Primary" key contains the subscription that is setup to receive messages from subgraph. 
"Rest" key contains the queries that willl be executed for each received message by the router.

Printing Query Plans was broken for PubSub subscriptions. 
From now on, for this type we return empty Fetch node for "Primary" key.