feat: validate optional required fields
This change enables validation of resolved input field set specified as
an argument for "requires" directive. If this field set 'points' to
a nullable fields marked as external, and if router resolves such
a field set with null values with errors pointing to those values, then it
will not provide such entities to the resolver of the field on which
directive was specified.