feat: validate optional "requires" fields (#2230)
This change enables validation of resolved input field set 
specified as an argument for "requires" directive.
If this field set 'points' to a nullable fields marked as external,
and if router resolves such a field set with null values 
with errors pointing to those values,
then it will not provide to the resolver entities that includes the field on 
which the "requires" directive was specified.

Use the `validate_required_external_fields` option to enable the feature.