fix: support requireFetchReasons on interface fields (#2272)
When an interface field is marked,
and a user requests that field in an operation, 
then fetch reasons for this field will be extended 
with fetch reasons for all the implementing types.

In general, a marked interface field leads to all 
implementation's fields being used for propagation.

I have removed special config needed for testing fetch reasons.
Instead I added those directives to the demo package.

Bumped the shoenig/go-m1cpu package to get rid of warnings from xcode 26.