fix: prevent potential race condition
The updateSubscription method is running once per subscription client in seperate go routines. The events slice passed into this function is shared between all routines. The hook execution in line 107 can either return the original events slice or a new deep copy, if events go modified by a hook.

After a hook call returned we cleaned up the events slice from nil elements, thus modifying it. If the hook call prior to that returned the original events slice instead of a deep copy, we caused a race condition, since we operate on the original slice with its underlying shared array.

To prevent this I removed the cleanup of nil elements and instead filter for these elements on the only place where they could hurt: The subscription update call to the engine. This way the updateSubscription function does not write to the slice at all, only reads. The only writes could happen in the hook call and our hook interface design should prevent modifying the original slice and instead of to do a deep copy first.