fix: remove unnecessary semaphore channel close
It might not be garantueed that there are no more subscription updaters running when the event updater Close() function is called. There is no logic waiting for the channel to close and since the garbage collector cleans up orphaned channels, there is no actual need to close the channel here.