fix: escape input when marshalling json template
The MarshalJSONTemplate method creates json-like output data, but it's not actually json conform. It contains placeholders with dollar signs as delimiters. For this reason we cannot use a json marshaller to render the output and resorted to creating the string manually. This could become a security problem because we do not really validate the input we use to create the output. Hypothetically an attacker could escape in that function and inject custom fields into the output. To circumvent this we marshall as many fields as we can. The only remaining field p.Event.Data does not contain user-input data, so we are safe not escaping it.