chore: reintroduce StreamHandlerError
core.NewHttpGraphqlError() is not really the best way to describe
an error in hooks processing. We have to set http specific codes
and status texts, which does not really fit well with subscriptions.
I created a new error for this called StreamHookError, which lets you
pass a message. It can be returned from a hook:

  return core.StreamHandlerError{Message: "my hook error"}

and it gets sent to the subscription client

  {"id":"1","type":"error","payload":[{"message":"my hook error"}]}

Afterwards we close the connection. This behaviour remains unchanged.