fix: remove unwanted logging of errors from OnPublishEvent hook
If this hook returns an error it is not wanted that the error is logged.
Other Cosmo Streams hooks don't do it. The reason is we want the hook developer
to decide what log level to use for errors, or if he doesn't want to log an error at all.
This error log handling is already in place for the other two hooks but somehow
was forgotten on this one.